started working a bit on this, not done, but working
added .idea files to gitigore and are working on the java badge more seriously

got the Java function to force an out of memory exception

the 10k sorted values are done correctly

had an issue, probably still there, but is nicer

added the enum class

just a few little changes

more stuff on inheritance stuff

got static stuff done

got the pass by reference done

added the .gitignore helper

fixed the error exceptions. I've got it done

added the virtualMethod stuff that is now finished

the instanceof file is done

the main issue is the final values, not sure how to fix that just yet

small changes

has a logger installed. still working on it

updated stuff, added a get method that essentially does nothing

maybe nicer?

looks a lot nicer now

added a logger

got a bit of the *.error working

working still

the project is done

more work

almost ready for a code review

got the compositionInheritance a lot nicer

added more stuff

should be ready for code review on this

quite a bit of refactoring to get it to work

almost ready for a review of the badge work

still more work